+++ examples/example.rs
@@ -46,9 +46,9 @@ impl ApplicationHandler for App {
         if let WindowEvent::CloseRequested = &event { event_loop.exit() }
     }
     fn device_event(
-        &mut self, _: &ActiveEventLoop, _: DeviceId, event: DeviceEvent
+        &mut self, _: &ActiveEventLoop, id: DeviceId, event: DeviceEvent
     ) {
-        self.input.update_with_device_event(&event);
+        self.input.update_with_device_event(id, &event);
     }
     fn about_to_wait(&mut self, _: &ActiveEventLoop) {
         self.input.update_with_gilrs(&mut self.gilrs);